Spotify_Recommendation_System/requirements.txt
flask
flask_cors
numpy
pandas
faiss-cpu
scikit-learn
sentence-transformers
gunicorn    # optional for production
